Drop min-edits value
It's quite common during a 24 hour period for an article to receive
less than 30 edits but to be one of the top edits for a day.

Min speed cut off is 0.1
30/300 = 0.1 so this means anything that doesn't reach 30 edits will
be lost after 5hrs.

For example Donald Harvey got 27 edits today by 17 editors and 6 anons
but was lost after trending for 6 hours.

The combination with min_editors of 3 means the bar is still high
enough to avoid the cache getting too big.

Bug: T160127
Change-Id: I78e941bb434921886612ff4565c04399eedc1912